+++ R/validate_estim_opt.R
@@ -43,7 +43,7 @@ validate_estim_opt <- function(estim_opt_input) {
   }
   
   # Check whether the optimizer can use the specified method
-  if (tolower(estim_opt[["optimizer"]]) %in% c("maxlik", "nloptr", "ucminf")) {
+  if (tolower(estim_opt[["optimizer"]]) %in% c("maxlik", "ucminf")) {
     
     # Check maxlik methods ----
     if (tolower(estim_opt[["optimizer"]]) == "maxlik") {
@@ -55,18 +55,6 @@ validate_estim_opt <- function(estim_opt_input) {
       } else {
         message(red$bold(symbol$cross), "  estim_opt().\n")
         stop("The optimization method must be either 'bfgs', 'bhhh' or 'nr' when using 'maxlik'. See ?maxLik")
-      }
-    }
-    
-    # Check nloptr options ----
-    if (tolower(estim_opt[["optimizer"]]) == "nloptr") {
-      if (tolower(estim_opt[["method"]]) %in% c("bfgs", "slsqp", "sbplx")) {
-        if (tolower(estim_opt[["method"]]) == "bfgs") estim_opt[["method"]] <- "NLOPT_LD_LBFGS"
-        if (tolower(estim_opt[["method"]]) == "slsqp") estim_opt[["method"]] <- "NLOPT_LD_SLSQP"
-        if (tolower(estim_opt[["method"]]) == "sbplx") estim_opt[["method"]] <- "NLOPT_LN_SBPLX"
-      } else {
-        message(red$bold(symbol$cross), "  estim_opt().\n")
-        stop("The optimization method must be either 'bfgs', 'slsqp' or 'sbplx' when using 'nloptr'. See ?nloptr")
       }
     }
     
@@ -82,7 +70,7 @@ validate_estim_opt <- function(estim_opt_input) {
     }
   } else {
     message(red$bold(symbol$cross), "  estim_opt().\n")
-    stop("The optimizer is either 'maxlik', 'nloptr' or 'ucminf'. See package documentation for more details.")
+    stop("The optimizer is either 'maxlik' or 'ucminf'. See package documentation for more details.")
   }
   
   # Check that enough cores are available ----